requirements.txt
flask>=2.0.0
sentence-transformers>=2.2.0
langchain>=0.1.0
faiss-cpu>=1.7.0
pandas>=1.3.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0